refactor(plugin-input-exercise): extract evaluation code
