+++ packages/editor/src/plugins/input-exercise/renderer.tsx
@@ -3,22 +3,25 @@ import { useInstanceData } from '@serlo/frontend/src/contexts/instance-context'
 import { cn } from '@serlo/frontend/src/helper/cn'
 import { useEffect, useState } from 'react'
 
+import { getCorrectAnswers } from './helper/get-correct-answers'
 import { InputExerciseType } from './input-exercise-type'
 
-type MathjsImport = typeof import('mathjs')
+export type MathjsImport = typeof import('mathjs')
+
+export interface InputExerciseAnswer {
+  value: string
+  isCorrect: boolean
+  feedback: JSX.Element | null
+}
 
 interface InputExersiseRendererProps {
-  type: string
+  type: InputExerciseType
   unit: string
-  answers: {
-    value: string
-    isCorrect: boolean
-    feedback: JSX.Element | null
-  }[]
+  answers: InputExerciseAnswer[]
   onEvaluate?: (correct: boolean, val: string) => void
 }
 
-interface FeedbackData {
+export interface FeedbackData {
   correct: boolean
   message: JSX.Element
 }
@@ -33,14 +36,25 @@ export function InputExerciseRenderer({
   const [value, setValue] = useState('')
   const exStrings = useInstanceData().strings.content.exercises
 
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
   const [mathjs, setMathjs] = useState<MathjsImport | null>(null)
   useEffect(() => void import('mathjs').then((math) => setMathjs(math)), [])
 
   function handleEvaluate() {
-    const feedbackData = checkAnswer()
-    if (onEvaluate) onEvaluate(feedbackData.correct, value)
-    setFeedback(feedbackData)
+    if (!mathjs) return
+
+    const correctAnswers = getCorrectAnswers(answers, value, type, mathjs)
+
+    const hasCorrectAnswer =
+      !!correctAnswers.length && correctAnswers[0].isCorrect
+    const customFeedbackNode = correctAnswers[0]?.feedback ?? null
+
+    if (onEvaluate) onEvaluate(hasCorrectAnswer, value)
+    setFeedback({
+      correct: hasCorrectAnswer,
+      message: customFeedbackNode ?? (
+        <>{exStrings[hasCorrectAnswer ? 'correct' : 'wrong']}</>
+      ),
+    })
   }
 
   return (
@@ -84,57 +98,4 @@ export function InputExerciseRenderer({
       </div>
     </div>
   )
-
-  function checkAnswer(): FeedbackData {
-    const filteredAnswers = answers.filter((answer) => {
-      try {
-        const solution = normalize(answer.value)
-        const submission = normalize(value)
-        if (!solution || !submission) return false
-
-        if (
-          type === 'input-expression-equal-match-challenge' &&
-          typeof solution === 'number' &&
-          typeof submission === 'number'
-        ) {
-          return solution - submission === 0
-        }
-        return solution === submission
-      } catch (e) {
-        return false
-      }
-    })
-    const customFeedbackNode = filteredAnswers[0]?.feedback ?? null
-
-    const hasCorrectAnswer =
-      filteredAnswers.length > 0 && filteredAnswers[0].isCorrect
-
-    return {
-      correct: hasCorrectAnswer,
-      message: customFeedbackNode ?? (
-        <>{exStrings[hasCorrectAnswer ? 'correct' : 'wrong']}</>
-      ),
-    }
-  }
-
-  function normalize(value: string) {
-    const _value = collapseWhitespace(value)
-    switch (type) {
-      case InputExerciseType.NumberExact:
-        return normalizeNumber(_value).replace(/\s/g, '')
-      case InputExerciseType.ExpressionEqual:
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-call, @typescript-eslint/no-unsafe-member-access
-        return Number(mathjs?.evaluate(normalizeNumber(_value)))
-      case InputExerciseType.StringNormalized:
-        return _value.toUpperCase()
-    }
-  }
-
-  function collapseWhitespace(val: string): string {
-    return val.replace(/[\s\xa0]+/g, ' ').trim()
-  }
-
-  function normalizeNumber(val: string) {
-    return val.replace(/,/g, '.').replace(/^[+]/, '')
-  }
 }
